requirements.txt
autopep8==1.7.0
Flask==2.2.2
Flask-SocketIO==5.3.1
Flask-SQLAlchemy==3.0.2
mypy==0.982
pylint==2.15.4
